clean up and reorganize files into different classes
